# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Serviço Backend",
        "description": "API RESTful para gerenciamento de serviços comunitários, incluindo pessoas, perfis, missões, notícias e doações.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/doacoes/{doacaoId}/itens": {
            "get": {
                "tags": [
                    "Itens de Doação"
                ],
                "summary": "Lista itens de uma doação",
                "operationId": "9022b53db5e256f627253afab1414806",
                "parameters": [
                    {
                        "name": "doacaoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de itens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/itens-doacao/{id}": {
            "get": {
                "tags": [
                    "Itens de Doação"
                ],
                "summary": "Busca item por ID",
                "operationId": "e733cd28bd6340fa4f74738e62934af6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Itens de Doação"
                ],
                "summary": "Atualiza um item de doação",
                "operationId": "b47b622b21dff6432c59a703bad0d91d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tipo_item_id": {
                                        "type": "integer"
                                    },
                                    "quantidade": {
                                        "type": "number"
                                    },
                                    "obs": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item não encontrado"
                    },
                    "409": {
                        "description": "Não é possível alterar, doação já foi entregue"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Itens de Doação"
                ],
                "summary": "Exclui um item de doação",
                "operationId": "c32259d1b7755b5a0dc279a35cbd65b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item excluído"
                    },
                    "404": {
                        "description": "Item não encontrado"
                    },
                    "409": {
                        "description": "Não é possível excluir, doação já foi entregue"
                    }
                }
            }
        },
        "/api/itens-doacao": {
            "post": {
                "tags": [
                    "Itens de Doação"
                ],
                "summary": "Cria um novo item de doação",
                "operationId": "4270730590bd3ef2c89f37f1f8aa9e57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "doacao_id",
                                    "tipo_item_id",
                                    "quantidade"
                                ],
                                "properties": {
                                    "doacao_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_item_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantidade": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5
                                    },
                                    "obs": {
                                        "type": "string",
                                        "example": "Observação do item"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
...
    },
...
}